feat: enemies are more likely to spawn in the direction the player is moving

+++ Assets/Scripts/Enemy/SOWave.cs
@@ -20,5 +20,5 @@ public class SOWave : ScriptableObject
 
     public float WaveTime() => _maxEnemyCount * _spawnInterval;
 
-    public static int CalcSpawnCost(EnemyAI enemy, int count) => enemy.spawnCost * count;
+    public static int CalcSpawnCost(EnemyAI enemy, int count) => (int)enemy.spawnCost * count;
 }
